# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RideConnect API
  description: |
    # RideConnect Backend API Documentation

    ## Overview
    RideConnect is a comprehensive ride-sharing and booking platform that connects passengers with drivers.
    This API provides secure access to all mobile app and backend management features.

    ## Authentication
    - **Mobile Users (Passengers/Drivers)**: Use Laravel Sanctum token-based authentication
    - **Managers (Admin/Accountant/Officer)**: Use token-based authentication with role-based access control

    ## Base URL
    ```
    https://api.rideconnect.com/api/v1
    ```

    ## Rate Limiting
    - 100 requests per minute for authenticated users
    - 20 requests per minute for unauthenticated users

    ## Response Format
    All responses follow a consistent JSON structure:
    ```json
    {
      "success": true,
      "message": "Success message",
      "data": { ... }
    }
    ```

    ## Error Handling
    - 400: Bad Request - Invalid input
    - 401: Unauthorized - Invalid or missing authentication
    - 403: Forbidden - Insufficient permissions
    - 404: Not Found - Resource not found
    - 422: Validation Error - Invalid request data
    - 500: Server Error - Internal server error

  version: 1.0.0
  contact:
    name: RideConnect API Support
    email: api-support@rideconnect.com
  license:
    name: Proprietary
servers:
  - url: https://api.rideconnect.com/api/v1
    description: Production Server
...
security:
  - BearerAuth: []
paths:
...
  /rider/documents:
    post:
      tags:
        - Rider
      summary: Upload document
      description: Upload driver document (license, insurance, etc.)
      operationId: uploadDocument
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - document_type
                - document
              properties:
                document_type:
                  type: string
                  enum: [license, insurance, registration, id_card, other]
                document:
                  type: string
                  format: binary
                expiry_date:
                  type: string
                  format: date
      responses:
        "201":
          description: Document uploaded
    get:
      tags:
        - Rider
      summary: List documents
      description: Get list of uploaded documents
      operationId: listDocuments
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Documents listed

  # ===========================
  # PAYMENTS
  # ===========================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT tokens are issued upon successful login.
        Include the token in the Authorization header:
        `Authorization: Bearer <token>`

  # ===========================
  # PARAMETERS
  # ===========================
...